tests/submission/Infraware/Forms/contents/Forms/output_name.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>output_name</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
	 <output name=o id='output_id'></output>
	</form>

	<script>
		
		var output = document.getElementById("output_id");
		name = output.name;
		output.name = "output_name";

		if (typeof(output.name) == "string") {
			if (name == 'o' && output.name == "output_name") {
				test(function() {
					assert_true(true, "name attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("name attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("name attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
